remotive_car/instances/android/pyproject.toml
[project]
name = "remotivelabs-topology-android-example"
description = "RemotiveTopology android example"
version = "1.0.0"

requires-python = ">=3.10,<4.0"

dependencies = [
  "structlog==25.4.0",
  "pytest==8.4.2",
  "pytest-asyncio==1.0.0",
  "transitions==0.9.2",
  "jupyterlab >=4.3.4",
  "ipywidgets >=8.1.5",
  "pillow >=11.1.0",
  "gherkin-official >=30.0.4",
  "behave >=1.2.6",
  "remotivelabs-broker~=0.9.1",
  "remotivelabs-topology~=0.9.4",
]


[dependency-groups]
dev = [
  "poethepoet>=0.18.0",
  "ruff>=0.11.10",
  "mypy>=1.14.1",
  "pytest>=8.4.2",
  "pytest-asyncio>=1.0.0",
]

[tool.uv.build-backend]
module-root = ""
module-name = ["cloudfeeder"]

[tool.poe.tasks]
format = [{ cmd = "ruff format ." }, { cmd = "ruff check --fix ." }]
ruff = [{ cmd = "ruff check ." }, { cmd = "ruff format --check --diff ." }]
mypy = [{ cmd = "mypy cloudfeeder" }]
lint = ["ruff", "mypy"]
test = { cmd = "echo no test in project" }

[tool.ruff]
extend = "../../../ruff.toml"

[tool.mypy]
python_version = "3.10"
disallow_untyped_calls = false
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
hide_error_codes = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
